Adjust CSS properties for more mobile friendliness
This makes a minimal set of changes to CSS properties for the pages to be more
mobile-friendly. Main changes:

1. Use more compact margins on narrow screens, leveraging a new `--margin`
variable and a width breakpoint.
2. Make navigation menu wrap.
3. Make baseline banner wrap when it's too long, dropping `width: max-content`
in particular.
4. Make baseline banner wrap below the title, switching from float to flexbox.

That's really meant to be a minimal set of changes to improve
mobile-friendliness (#5). The navigation menu could perhaps be turned into a
hamburger menu on narrow screens as it takes a bit of screen estate as things
stand. The statuses per browser could perhaps also benefit from being presented
differently, but they already wrap nicely thanks to flexbox. Any, lots of
further improvements possible!