Merge bump-deps and generate-site workflows
By default, a push triggered by a GitHub workflow cannot trigger another
workflow, see:
https://docs.github.com/en/actions/using-workflows/triggering-a-workflow#triggering-a-workflow-from-a-workflow

This makes it harder to have separate workflows to bump dependencies and
re-generate the site.

There are different ways to work around that, including:
1. Create and use a Personal Access Token instead of `GITHUB_TOKEN` to push.
This requires someone to create and maintain a Personal Access Token for the
job, which is doable but seems a bit overkill.
2. Add a `workflow_run` trigger in the workflow that generates the site to
make it run when the bump-deps workflow is done. That works well, but then
ideally, site generation would only run if dependencies were effectively bumped
and the canonical way to pass data from one workflow to another is through
artifacts, which also seems overkill, see:
https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#using-data-from-the-triggering-workflow
3. Merge the workflows and use a little bit of conditional logic to distinguish
the cases "triggered by a push" from "triggered by cron".

This update implements 3. The workflow may also be triggered manually, in which
case it will also tries to bump dependencies.

Still to be done (separately):
- The code that bumps dependencies should not commit anything if there was no
update.
- The site should probably not be generated if there was no update.