Add timeline charts to Timeline page
Inspired by https://github.com/tidoust/web-features-viz/ and discussions within
the WebDX Community Group and #20, this adds 3 charts to the Timeline page to
visualize the evolution of the web platform:

- A chart that shows the evolution of the web platform in terms of number of
features per Baseline status.
- A chart that shows the same thing with percentages to illustrate the
evolution of the distribution of features between the different Baseline
statuses.
- A box plot chart to show the evolution of the duration it takes for a feature
to go from first implementation available to Baseline Newly Available.

The charts are introduced by a bit of text to ease reading.

The charters are rendered thanks to the open-source ApexCharts library. They
are interactive.

The page fetches the data from JSON files that get updated through the script
`additional-data/scripts/update-timeline-stats.js`, run as part of the job that
updates the site (I did not create a separate job because stats should ideally
be updated at the same time as web-features gets updated).

The JSON files are copied to the site so that users can also download them if
they so wish.

The ApexCharts library, like most interactive chartering libraries, is a bit
heavy (>500KB minified) but only loaded for that page. The additional CSS is
small and loaded everywhere for simplicity.