Add timeline charts to Timeline page (#53)
Inspired by https://github.com/tidoust/web-features-viz/ and discussions within
the WebDX Community Group and #20, this adds 3 charts to the Timeline page to
visualize the evolution of the web platform:

- A chart that shows the evolution of the web platform in terms of number of
features per Baseline status.
- A chart that shows the same thing with percentages to illustrate the
evolution of the distribution of features between the different Baseline
statuses.
- A box plot chart to show the evolution of the duration it takes for a feature
to go from first implementation available to Baseline Newly Available.

The charts are introduced by a bit of text to ease reading.

The charters are rendered thanks to the open-source ApexCharts library. They
are interactive. ApexCharts is added as a dev dependency.

The page fetches the data from JSON files that get updated through the script
`update-timeline-stats.js`, run as part of the job that updates the site (I did not
create a separate job because stats should ideally be updated at the same time
as web-features gets updated).

The JSON files are copied to the site so that users can also download them if
they so wish.

The ApexCharts library, like most interactive chartering libraries, is a bit
heavy (>500KB minified) but only loaded for that page.

Styles of the timeline table were also updated: a black background appeared
when the window was large enough to contain the whole table. Also, with the
charts added at the top of the page, it probably no longer makes sense to
create inner vertical scrolling for the table.